main: Add quick-n-dirty interactive tutorial for the first level
Have the game show a pre-defined message on reaching the goal in the
first level.  Most players will probably figure out that they should
get the gold coin.  Those will probably also figure out that the arrow
keys (or "wasd") will help them get there.

An experienced player may from here guess (from the lack of reaction),
that he/she should return to start.  But unless you know that there is
a separate key for entering the time machine, you are pretty much
f***ed from there.

Of course, you could look up that key in the ingame help menu.  But
most players are just going to assume the game is broken.  So, spam
the player with a helpful little message in the first level.

Signed-off-by: Niels Thykier <niels@thykier.net>